# backend/internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Songguo Admin API
  version: "1"
  description: >
    Control-plane (admin) API for the Songguo AI usage gateway. It is read-mostly
    and consumed by the dashboard, the `songguo` MCP server, and any other tooling.
    Every endpoint is gated by a single admin bearer key (`SONGGUO_ADMIN_KEY`);
    when that key is empty the server runs the API unprotected and logs a warning.

    This describes only the control plane. The data plane — the transparent proxy
    at the native vendor paths (`/v1/*`, `/api/v3/*`), with the provider selected
    by the `X-Songguo-Provider` header, else the body's model, else the wire
    default — speaks each vendor's native wire protocol and is documented in
    docs/registry.md.
servers:
  - url: /
security:
  - adminKey: []
...
paths:
...
  /api/users:
    get:
      tags: [users]
      summary: List users with computed lifetime spend
      operationId: listUsers
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/User" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [users]
      summary: Create a user (plaintext key returned once)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUser" }
      responses:
        "201":
          description: The created user, including the plaintext `key`
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/users/{id}:
    get:
      tags: [users]
      summary: Get one user with computed lifetime spend
      operationId: getUser
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [users]
      summary: Update a user's mutable fields
      operationId: patchUser
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchUser" }
      responses:
        "200":
          description: The updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [users]
      summary: Delete a user (the admin user cannot be deleted)
      operationId: deleteUser
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: User deleted }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/users/{id}/revoke:
    post:
      tags: [users]
      summary: Revoke a user's key
      operationId: revokeUser
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The revoked user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    adminKey:
      type: http
      scheme: bearer
      description: The admin key from SONGGUO_ADMIN_KEY.
...
  responses:
    Unauthorized:
      description: Missing or invalid admin key
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            type: { type: string, description: "songguo_-prefixed reason, e.g. songguo_not_found" }
...
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        key_prefix: { type: string }
        budget: { type: [number, "null"] }
        scope: { type: array, items: { type: string } }
        rpm: { type: integer }
        capture: { type: [boolean, "null"] }
        created_at: { type: string, format: date-time }
        revoked_at: { type: [string, "null"], format: date-time }
        spent: { type: number }
        active: { type: boolean }
        last_seen: { type: [string, "null"], format: date-time, description: "Timestamp of the user's most recent call, or null if never used." }
        key: { type: string, description: "Plaintext key. Empty for users created before key storage existed." }
    CreateUser:
      type: object
      required: [name]
      properties:
        name: { type: string }
        budget: { type: number, description: "USD budget cap." }
        scope: { type: array, items: { type: string }, description: "Allowed model strings." }
        rpm: { type: integer }
        capture: { type: boolean }
    PatchUser:
      type: object
      properties:
        name: { type: string }
        budget: { type: number }
        scope: { type: array, items: { type: string } }
        rpm: { type: integer }
        capture: { type: boolean }
...